Added additional debug statements and handled corner cases
I added some additional debugging information so that it is more clear
what exactly is happening. I also fixed deadzone issues with the turning
so that you have to be a little more deliberate to move past center.  I
also added in the ability to see what the max speed is when you adjust
it